Replace capabilities with extraCapabilities.
The relationship of each properties in capabilities is OR. When simulcast
is enabled, transcoding will be disabled. In this case, capabilities cannot
represent a stream's settings, which should be included in capabilites.
extraCapabilities doesn't include settings, it only includes transcoding
formats for now. "Transcoding formats" means the formats described in
optional property of a stream.