Refactor p2p sample (#383)
* Move JS out of peercall.html and run js-beautify on sample

* Fix most ESLint issues on P2P sample

Also, move user-editable data to top of peercall.js and update README.

Add an .eslintrc to samples directory that knows we're developing code
for the browser, etc.

A few ESLint errors related to undefined variables and indenting that
js-beautify does wrong are left.

* Clarify steps in P2P sample README

If someone digs down into the tree and looks at this sample, they may not know
how to get owt.js. That's what happened to me the first time I used this.