Setting local description is considered as negotiating.
When the impolite side receives offer after `setLocalDescription`, before it's
resolved, it sets both local description and remote description. Its
signaling state changes from stable to have-local-offer, to stable, to
have-remote-offer. It doesn't comply with the behavior of impolite side
described in perfect negotiation, which is ignoring the remote
description.

Test: P2P case WebRTC collision should be resolved.