Create PeerConnection when PeerConnectionChannel is created. (#532)
* Create PeerConnection when PeerConnectionChannel is created.

This change allows application to get PeerConnection before publishing
or subscribing. Then the application can create RTCRtpTransceivers on
the PeerConnection returned, and publish a stream with
RTCRtpTransceivers pre-configurated.

* Notify ConferenceClient when a new publication ID is ready.

This is actually not needed because conference server supports multiple
streams on a single PeerConnection, so only a single
PeerConnectionChannel is created. But let's keep the old behavior until
ConferenceClient is updated.